+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to the "lazykey" extension will be documented in this file.
 
 ## [v0.0.4]
+
+- Add auto generate variable name
 - Fix `0` can't be entered in some cases